+++ BattleBones/Assets/Scripts/ScriptableObjects/FieldType.cs
@@ -1,12 +1,10 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 [CreateAssetMenu(fileName = "New FieldType", menuName = "FieldType")]
 public class FieldType : ScriptableObject
 {
     public string fieldName;
-    public float movementPointsCost;
+    public int movementPointsCost;
     public bool isObstacle;
     public bool isBlockSight;
 }
